requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via supermemo2
backcall==0.2.0
    # via ipython
bleach==6.0.0
    # via readme-renderer
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   click-log
    #   python-semantic-release
click-log==0.4.0
    # via python-semantic-release
colorama==0.4.6
    # via twine
comm==0.1.3
    # via ipykernel
cryptography==40.0.1
    # via
    #   pyjwt
    #   secretstorage
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via python3-openid
deprecation==2.1.0
    # via postgrest
dj-database-url==1.3.0
    # via django-heroku
django==4.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-allauth
    #   django-extensions
    #   django-heroku
    #   djangorestframework
django-allauth==0.54.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-heroku==0.3.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
docutils==0.19
    # via readme-renderer
dotty-dict==1.3.1
    # via python-semantic-release
executing==1.2.0
    # via stack-data
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via python-semantic-release
gotrue==1.0.1
    # via supabase
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
importlib-metadata==6.3.0
    # via
    #   keyring
    #   twine
invoke==1.7.3
    # via python-semantic-release
ipykernel==6.22.0
    # via ipywidgets
ipython==8.12.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.6
    # via -r requirements.in
jaraco-classes==3.2.3
    # via keyring
jedi==0.18.2
    # via ipython
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jupyter-client==8.2.0
    # via ipykernel
jupyter-core==5.3.0
    # via
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.7
    # via ipywidgets
keyring==23.13.1
    # via twine
markdown==3.4.3
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
more-itertools==9.1.0
    # via jaraco-classes
nest-asyncio==1.5.6
    # via ipykernel
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   deprecation
    #   ipykernel
    #   python-semantic-release
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pkginfo==1.9.6
    # via twine
platformdirs==3.2.0
    # via jupyter-core
postgrest==0.10.6
    # via supabase
prompt-toolkit==3.0.38
    # via ipython
psutil==5.9.4
    # via ipykernel
psycopg2==2.9.6
    # via
    #   -r requirements.in
    #   django-heroku
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   gotrue
    #   postgrest
pygments==2.15.0
    # via
    #   ipython
    #   readme-renderer
pyjwt[crypto]==2.6.0
    # via django-allauth
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   realtime
    #   storage3
python-gitlab==3.14.0
    # via python-semantic-release
python-semantic-release==7.33.2
    # via supabase
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via djangorestframework
pyzmq==25.0.2
    # via
    #   ipykernel
    #   jupyter-client
readme-renderer==37.3
    # via twine
realtime==1.0.0
    # via supabase
requests==2.28.2
    # via
    #   django-allauth
    #   python-gitlab
    #   python-semantic-release
    #   requests-oauthlib
    #   requests-toolbelt
    #   twine
requests-oauthlib==1.3.1
    # via django-allauth
requests-toolbelt==0.10.1
    # via
    #   python-gitlab
    #   twine
rfc3986[idna2008]==1.5.0
    # via
    #   httpx
    #   twine
secretstorage==3.3.3
    # via keyring
semver==2.13.0
    # via python-semantic-release
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlparse==0.4.3
    # via django
stack-data==0.6.2
    # via ipython
storage3==0.5.2
    # via supabase
strenum==0.4.10
    # via postgrest
supabase==1.0.3
    # via -r requirements.in
supafunc==0.2.2
    # via supabase
supermemo2==2.0.0
    # via -r requirements.in
tomlkit==0.11.7
    # via python-semantic-release
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.65.0
    # via
    #   -r requirements.in
    #   twine
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
twine==3.8.0
    # via python-semantic-release
typing-extensions==4.5.0
    # via
    #   dj-database-url
    #   pydantic
    #   realtime
    #   storage3
urllib3==1.26.15
    # via
    #   requests
    #   twine
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websockets==10.4
    # via realtime
wheel==0.40.0
    # via python-semantic-release
whitenoise==6.4.0
    # via
    #   -r requirements.in
    #   django-heroku
widgetsnbextension==4.0.7
    # via ipywidgets
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
